refactor memory endpoints in multiple files
