+++ core/cat/main.py
@@ -19,11 +19,11 @@ from cat.routes import (
     llm,
     embedder,
     auth_handler,
-    memory,
     plugins,
     upload,
     websocket,
 )
+from cat.routes.memory.memory_router import memory_router
 from cat.routes.static import admin, static
 from cat.routes.openapi import get_openapi_configuration_function
 from cat.looking_glass.cheshire_cat import CheshireCat
@@ -79,7 +79,7 @@ cheshire_cat_api.add_middleware(
 )
 
 # Add routers to the middleware stack.
-cheshire_cat_api.include_router(base.router, tags=["Status"])
+cheshire_cat_api.include_router(base.router, tags=["Home"])
 cheshire_cat_api.include_router(auth.router, tags=["User Auth"], prefix="/auth")
 cheshire_cat_api.include_router(users.router, tags=["Users"], prefix="/users")
 cheshire_cat_api.include_router(settings.router, tags=["Settings"], prefix="/settings")
@@ -88,7 +88,7 @@ cheshire_cat_api.include_router(
 )
 cheshire_cat_api.include_router(embedder.router, tags=["Embedder"], prefix="/embedder")
 cheshire_cat_api.include_router(plugins.router, tags=["Plugins"], prefix="/plugins")
-cheshire_cat_api.include_router(memory.router, tags=["Memory"], prefix="/memory")
+cheshire_cat_api.include_router(memory_router, prefix="/memory")
 cheshire_cat_api.include_router(
     upload.router, tags=["Rabbit Hole"], prefix="/rabbithole"
 )
